time syscall implemented

+++ src/arch/riscv/sbi/common.c
@@ -5,4 +5,4 @@ void port_write_ch(char ch) { sbi_console_putchar((int)ch); }
 
 void port_write(char *str) { sbi_console_putstr(str); }
 
-int sys_port_read() { return sbi_console_getchar(); }
+long sys_port_read() { return sbi_console_getchar(); }
